+++ lib/book/book_detail.dart
